Fix my bad use of dist in collision checking
'dist' was the absolute distance /along the current axis/, so while
comparing it to zero was always kosher, comparing different dists made
no sense whatsoever.

The thing I actually wanted to measure was how much of the movement was
allowed, so now I just do that directly.  As a huge bonus, that lets me
know exactly which normals I'm colliding with, so I could get rid of a
massive ugly loop that tried to figure it out after the fact.