Fix a rather significant memory leak
Every time a map loads, it adds itself to the collider.  The collider
holds weak references to shapes (since actors, at least, hold onto their
shapes), so the map crammed all of its shapes into a single table.

The next time the map is loaded, the same map object is reused, but all
of its shapes are recreated and crammed into the same table, which still
contains all the old shapes...  etc.

Fix is to just store the shapes in such a way that they can be reused
next time.  Which should also make map loading slightly faster the
second time.